% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2mat.R
\name{dat2mat}
\alias{dat2mat}
\title{Dataframe to a matrix}
\usage{
dat2mat(dat, xcol, ycol, zcol)
}
\arguments{
\item{dat}{A dataframe, which is to be converted to a matrix.}

\item{xcol}{A string, which must be one of the column names of the dataframe given by "dat". This column vector will be used as positions along x-axis of the result matrix.}

\item{ycol}{A string, which must be one of the column names of the dataframe given by "dat". This column vector will be used as positions along y-axis of the result matrix.}

\item{zcol}{A string, which must be one of the column names of the dataframe given by "dat". This column vector will be used as z-values (colors) of the result matrix.}
}
\value{
A matrix.
}
\description{
This function converts a dataframe to a matrix. X-, y-, and z-axes can be given as arguments.
}
\examples{
dat = expand.grid('x'=1:4, 'y'=1:3)
dat$z = 1:nrow(dat)
mat = dat2mat(dat, xcol='x', ycol='y', zcol='z')
print(mat)
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
