% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2dat.R
\name{mat2dat}
\alias{mat2dat}
\title{Image matrix to a dataframe}
\usage{
mat2dat(mat)
}
\arguments{
\item{mat}{A 2-dimensional matrix to be converted to a dataframe.}
}
\value{
A dataframe, which has x, y, and z columns.
}
\description{
This function converts a 2-dimensional matrix to a dataframe.
}
\examples{
mat = matrix(1:12, nrow=3, byrow=TRUE)
dat = mat2dat(mat)
print(dat)
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
