% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycontour.R
\name{mycontour}
\alias{mycontour}
\title{Filled contour plot based on ggplot}
\usage{
mycontour(plt, ylim = NULL, zlim = NULL, breaks = 10, line_color_ratio = 4)
}
\arguments{
\item{plt}{A ggplot object. It should already have the input dataframe and the aes values for x, y, and z.}

\item{ylim}{A 2 value vector.}

\item{zlim}{A 2 value vector.}

\item{breaks}{An integer for the number of contour lines. The first and last values will be the ends of the zlim scale. Therefore, (breaks-2) contour lines will be visible in the result plot.}
}
\value{
A ggplot object, which can displays a basic filled contour with contour lines and labels.
}
\description{
This function renders a ggplot object to be able to produce a filled contour plot with contour lines and labels. (NOTE: the values used for the x and y axes must form a grid for the current version of this function.)
}
\examples{
set.seed(10)
dat = mgcv::gamSim()
mdl = mgcv::gam(y ~ te(x0,x1), data=dat)
vclist = list(x0=list(edge=0,len.out=100), x1=list(edge=0,len.out=100))
dat = create_ndat(mdl$model, vclist)
dat$fit = as.vector(mgcv::predict.gam(mdl, newdata=dat))
plt = ggplot2::ggplot(dat, aes(x=x0, y=x1, z=fit))
plt = mycontour(plt)
print(plt)
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
