% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_mat.R
\name{norm_mat}
\alias{norm_mat}
\title{Normalize values in a matrix}
\usage{
norm_mat(mat, rng = c(0, 255))
}
\arguments{
\item{mat}{A matrix, whose values will be normalized.}

\item{rng}{A range of normalized values, which is c(0,255) as default.}
}
\value{
A matrix whose values are normalized to a given range.
}
\description{
This function normalizes values in a matrix within a certain range (default 0-255).
}
\examples{
mat0 = matrix(seq(-2,9,by=1), nrow=4)
mat1 = norm_mat(mat0)
mat2 = norm_mat(mat0, c(-1,1))
print(mat0)
print(mat1)
print(mat2)
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
