% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_index.R
\name{to_index}
\alias{to_index}
\title{Conversion of a continuous vector to a discrete vector}
\usage{
to_index(vec)
}
\arguments{
\item{vec}{A numeric vector. Its minimum value becomes 1. Its maximum value becomes the length of the vector.}
}
\value{
A numeric vector.
}
\description{
This function simply sorts elements in a vector and give indices to each element. E.g. c(0.3, -0.08, 4.2) --> c(2, 1, 3). Therefore, please note that intervals between elements are discarded.
}
\examples{
vec = c(0.3, -0.08, 4.2)
print(to_index(vec))
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
